+++ rollup.config.js
@@ -3,14 +3,19 @@ import svelte from "rollup-plugin-svelte";
 import svelteReadme from "svelte-readme";
 import pkg from "./package.json";
 
-const DEV = process.env.ROLLUP_WATCH;
-const BUNDLE = process.env.BUNDLE === "true";
-
 export default () => {
-  if (!BUNDLE) {
+  if (process.env.BUNDLE !== "true") {
     return svelteReadme({
-      minify: !DEV,
-      prefixUrl: "https://github.com/metonym/svelte-search/tree/master/",
+      style: `
+        [data-svelte-search] input {
+          width: 100%;
+          font-size: 1rem;
+          padding: .5rem;
+          margin: .5rem 0;
+          border: 1px solid #e0e0e0;
+          border-radius: 0.25rem;
+        }
+      `
     });
   }