+++ types/index.d.ts
@@ -1,1 +1,1 @@
-export { default } from "./Search";
+export { default as default } from "./Search";