+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0](https://github.com/metonym/svelte-search/releases/tag/v1.0.0) - 2020-12-31
+
+**Breaking Changes**
+
+- Combine `debounce` and `debounceValue`; default `debounce` value is `0`
+- Use value for `e.detail` for dispatched "type" event
+- Add `autofocus`, `ref` props
+- Remove `name`, `clear`, `focus` props
+- Use `SvelteComponentTyped` in TypeScript definitions
+
 ## [0.2.0](https://github.com/metonym/svelte-search/releases/tag/v0.2.0) - 2020-11-17
 
 - Add TypeScript definitions